requirements.txt
asgiref==3.5.2
autopep8==1.7.0
certifi==2022.9.24
charset-normalizer==2.1.1
distlib==0.3.4
dj-database-url==1.0.0
Django==4.0.4
django-colorfield==0.7.2
django-extensions==3.1.5
django-filter==22.1
django-heroku==0.3.1
django-jazzmin==2.5.0
djangorestframework==3.13.1
filelock==3.7.0
gunicorn==20.1.0
heroku==0.1.4
idna==3.4
mysqlclient==2.1.1
Pillow==9.1.1
platformdirs==2.5.2
psycopg2==2.9.5
pycodestyle==2.9.1
PyMySQL==1.0.2
python-dateutil==1.5
python-decouple==3.6
pytz==2022.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2022.1
urllib3==1.26.12
virtualenv==20.14.1
waitress==2.1.2
whitenoise==6.2.0
